requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.2.1
Django==1.11.9
django-extensions==1.9.9
django-jet==1.0.7
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
enum34==1.1.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
nltk==3.2.5
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
scandir==1.6
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
traitlets==4.3.2
typing==3.6.4
Unipath==1.1
wcwidth==0.1.7
django-session-security==2.6.5
django-tracking==0.4.1
